Consolidate jwt and auth modules into capabilities.rs and add conversion impls for capability types from input types
