+++ rust-executor/src/agent/mod.rs
@@ -1,2 +1,1 @@
-pub mod auth;
-pub mod jwt;
+pub mod capabilities;
