+++ rust-executor/src/js_core/jwt_extension.rs
@@ -1,6 +1,6 @@
 use std::borrow::Cow;
 use deno_core::{error::AnyError, include_js_files, op2, Extension, Op};
-use crate::agent::jwt::{AuthInfo, Claims};
+use crate::agent::capabilities::{AuthInfo, Claims};
 
 #[op2(async)]
 #[string]
@@ -10,13 +10,13 @@ async fn generate_jwt(
     #[smi] expiration_time: u64,
     #[serde] capabilities: AuthInfo,
 ) -> Result<String, AnyError> {
-    crate::agent::jwt::generate_jwt(issuer, audience, expiration_time, capabilities)
+    crate::agent::capabilities::generate_jwt(issuer, audience, expiration_time, capabilities)
 }
 
 #[op2(async)]
 #[serde]
 pub async fn verify_jwt(#[string] token: String) -> Result<Claims, AnyError> {
-    crate::agent::jwt::decode_jwt(token)
+    crate::agent::capabilities::decode_jwt(token)
 }
 
 pub fn build() -> Extension {
